georag/requirements.txt
pymupdf
ollama
sentence-transformers
faiss-cpu
numpy
